% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{block_to_embeddings_registry_filename}
\alias{block_to_embeddings_registry_filename}
\title{Generate embedding registry filename for a block}
\usage{
block_to_embeddings_registry_filename(year, block_lon, block_lat)
}
\arguments{
\item{year}{Integer year}

\item{block_lon}{Block longitude}

\item{block_lat}{Block latitude}
}
\value{
Character string filename
}
\description{
Generate embedding registry filename for a block
}
\keyword{internal}
