% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{block_to_landmasks_registry_filename}
\alias{block_to_landmasks_registry_filename}
\title{Generate landmasks registry filename for a block}
\usage{
block_to_landmasks_registry_filename(block_lon, block_lat)
}
\arguments{
\item{block_lon}{Block longitude}

\item{block_lat}{Block latitude}
}
\value{
Character string filename
}
\description{
Generate landmasks registry filename for a block
}
\keyword{internal}
