% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{blocks_in_bounds}
\alias{blocks_in_bounds}
\title{Get all blocks that intersect with a bounding box}
\usage{
blocks_in_bounds(bbox)
}
\arguments{
\item{bbox}{sf bbox object or named vector with xmin, ymin, xmax, ymax}
}
\value{
Data frame with block_lon and block_lat columns
}
\description{
Get all blocks that intersect with a bounding box
}
\keyword{internal}
