% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{calculate_bbox_from_file}
\alias{calculate_bbox_from_file}
\title{Calculate bounding box from file}
\usage{
calculate_bbox_from_file(filepath)
}
\arguments{
\item{filepath}{Path to file}
}
\value{
sf bbox object
}
\description{
Extract bounding box from a GeoTIFF or vector file.
}
