% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_file_hash}
\alias{calculate_file_hash}
\title{Calculate SHA256 hash of a file}
\usage{
calculate_file_hash(file_path)
}
\arguments{
\item{file_path}{Path to file}
}
\value{
Character string hash
}
\description{
Calculate SHA256 hash of a file
}
\keyword{internal}
