% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country.R
\name{CountryLookup}
\alias{CountryLookup}
\title{Country Lookup Class}
\description{
Country Lookup Class

Country Lookup Class
}
\details{
Provides country name to geometry and bounding box lookup using Natural Earth data.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cache_dir}}{Directory for caching country data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CountryLookup-new}{\code{CountryLookup$new()}}
\item \href{#method-CountryLookup-get_country_geometry}{\code{CountryLookup$get_country_geometry()}}
\item \href{#method-CountryLookup-get_country_bbox}{\code{CountryLookup$get_country_bbox()}}
\item \href{#method-CountryLookup-find_country_for_point}{\code{CountryLookup$find_country_for_point()}}
\item \href{#method-CountryLookup-list_countries}{\code{CountryLookup$list_countries()}}
\item \href{#method-CountryLookup-clone}{\code{CountryLookup$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CountryLookup-new"></a>}}
\if{latex}{\out{\hypertarget{method-CountryLookup-new}{}}}
\subsection{Method \code{new()}}{
Create a new CountryLookup object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CountryLookup$new(cache_dir = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cache_dir}}{Directory for caching (default: user cache)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new CountryLookup object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CountryLookup-get_country_geometry"></a>}}
\if{latex}{\out{\hypertarget{method-CountryLookup-get_country_geometry}{}}}
\subsection{Method \code{get_country_geometry()}}{
Get the geometry for a country
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CountryLookup$get_country_geometry(country_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{country_name}}{Country name (case-insensitive)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
sf geometry object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CountryLookup-get_country_bbox"></a>}}
\if{latex}{\out{\hypertarget{method-CountryLookup-get_country_bbox}{}}}
\subsection{Method \code{get_country_bbox()}}{
Get the bounding box for a country
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CountryLookup$get_country_bbox(country_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{country_name}}{Country name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
sf bbox object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CountryLookup-find_country_for_point"></a>}}
\if{latex}{\out{\hypertarget{method-CountryLookup-find_country_for_point}{}}}
\subsection{Method \code{find_country_for_point()}}{
Find which country contains a point
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CountryLookup$find_country_for_point(lon, lat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lon}}{Longitude}

\item{\code{lat}}{Latitude}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Country name or NA if not found
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CountryLookup-list_countries"></a>}}
\if{latex}{\out{\hypertarget{method-CountryLookup-list_countries}{}}}
\subsection{Method \code{list_countries()}}{
List all available countries
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CountryLookup$list_countries()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character vector of country names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CountryLookup-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CountryLookup-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CountryLookup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
