% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country.R
\name{find_country_for_point}
\alias{find_country_for_point}
\title{Find country for a point}
\usage{
find_country_for_point(lon, lat, cache_dir = NULL)
}
\arguments{
\item{lon}{Longitude}

\item{lat}{Latitude}

\item{cache_dir}{Optional cache directory}
}
\value{
Country name or NA
}
\description{
Find country for a point
}
