% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_bbox}
\alias{format_bbox}
\title{Format a bounding box for display}
\usage{
format_bbox(bbox)
}
\arguments{
\item{bbox}{Named vector or sf bbox}
}
\value{
Character string
}
\description{
Format a bounding box for display
}
\keyword{internal}
