% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_utm_epsg}
\alias{get_utm_epsg}
\title{Get UTM EPSG code for coordinates}
\usage{
get_utm_epsg(lon, lat)
}
\arguments{
\item{lon}{Longitude}

\item{lat}{Latitude}
}
\value{
Integer EPSG code
}
\description{
Get UTM EPSG code for coordinates
}
\keyword{internal}
