% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_utm_zone}
\alias{get_utm_zone}
\title{Get UTM zone for a longitude}
\usage{
get_utm_zone(lon)
}
\arguments{
\item{lon}{Longitude in decimal degrees}
}
\value{
Integer UTM zone number
}
\description{
Get UTM zone for a longitude
}
\keyword{internal}
