% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{mask_embeddings_to_region}
\alias{mask_embeddings_to_region}
\title{Apply Spatial Mask to Embeddings}
\usage{
mask_embeddings_to_region(embeddings, tiles, region)
}
\arguments{
\item{embeddings}{List of 3D arrays from tiles}

\item{tiles}{List of Tile objects}

\item{region}{sf object defining the mask region}
}
\value{
List of masked 3D arrays (pixels outside region set to NA)
}
\description{
Masks embeddings to only include pixels within a spatial region.
Useful for irregular regions from shapefiles.
}
\keyword{internal}
