% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_embedding_bands}
\alias{plot_embedding_bands}
\title{Plot embedding bands}
\usage{
plot_embedding_bands(tile, bands = 1:9, ncol = 3)
}
\arguments{
\item{tile}{Tile object}

\item{bands}{Bands to plot (default first 9)}

\item{ncol}{Number of columns in plot grid}
}
\value{
ggplot object
}
\description{
Create a visualization of selected embedding bands from a tile.
}
