% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_npy}
\alias{read_npy}
\title{Read NPY file}
\usage{
read_npy(path)
}
\arguments{
\item{path}{Path to .npy file}
}
\value{
Array with data
}
\description{
Simple reader for numpy .npy files (assumes C-order, little-endian).
}
\keyword{internal}
