% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile.R
\name{tile_from_geotiff}
\alias{tile_from_geotiff}
\title{Create Tile from GeoTIFF file}
\usage{
tile_from_geotiff(geotiff_path)
}
\arguments{
\item{geotiff_path}{Path to GeoTIFF file}
}
\value{
Tile object
}
\description{
Create Tile from GeoTIFF file
}
