% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile.R
\name{tile_from_npy}
\alias{tile_from_npy}
\title{Create Tile from NPY files}
\usage{
tile_from_npy(embedding_path, base_dir = NULL)
}
\arguments{
\item{embedding_path}{Path to embedding .npy file}

\item{base_dir}{Base directory containing the tile structure}
}
\value{
Tile object
}
\description{
Create Tile from NPY files
}
