% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile.R
\name{tile_from_zarr}
\alias{tile_from_zarr}
\title{Create Tile from Zarr archive}
\usage{
tile_from_zarr(zarr_path)
}
\arguments{
\item{zarr_path}{Path to Zarr directory}
}
\value{
Tile object
}
\description{
Create Tile from Zarr archive
}
