% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tile_to_bbox}
\alias{tile_to_bbox}
\title{Create sf bounding box from tile coordinates}
\usage{
tile_to_bbox(lon, lat)
}
\arguments{
\item{lon}{Tile center longitude}

\item{lat}{Tile center latitude}
}
\value{
sf bbox object
}
\description{
Create sf bounding box from tile coordinates
}
\keyword{internal}
