% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tile_to_geotiff_path}
\alias{tile_to_geotiff_path}
\title{Generate GeoTIFF path for a tile}
\usage{
tile_to_geotiff_path(lon, lat, year)
}
\arguments{
\item{lon}{Tile longitude}

\item{lat}{Tile latitude}

\item{year}{Integer year}
}
\value{
Character string path
}
\description{
Generate GeoTIFF path for a tile
}
\keyword{internal}
