% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tile_to_landmask_filename}
\alias{tile_to_landmask_filename}
\title{Generate landmask filename for a tile}
\usage{
tile_to_landmask_filename(lon, lat)
}
\arguments{
\item{lon}{Tile longitude}

\item{lat}{Tile latitude}
}
\value{
Character string filename
}
\description{
Generate landmask filename for a tile
}
\keyword{internal}
