% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tile_to_bounds}
\alias{tile_to_bounds}
\title{Get tile bounds from center coordinates}
\usage{
tile_to_bounds(lon, lat)
}
\arguments{
\item{lon}{Tile center longitude}

\item{lat}{Tile center latitude}
}
\value{
Named list with xmin, ymin, xmax, ymax
}
\description{
Get tile bounds from center coordinates
}
\keyword{internal}
